refactor: move types to Types.hs module
