+++ src/Data/Aeson/JSONPath/Parser.hs
@@ -1,68 +1,23 @@
-{-# LANGUAGE DeriveLift #-}
 {-# OPTIONS_GHC -Wno-unused-do-bind #-}
 module Data.Aeson.JSONPath.Parser
-  ( JSPQuery (..)
-  , JSPSegment (..)
-  , JSPChildSegment (..)
-  , JSPDescSegment (..)
-  , JSPSelector (..)
-  , JSPWildcardT (..)
-  , pJSPQuery
-  )
+  ( pJSPQuery )
   where
 
 import qualified Data.Text                      as T
 import qualified Text.ParserCombinators.Parsec  as P
 
+import Data.Aeson.JSONPath.Types     (JSPQuery (..)
+                                     , JSPSegment (..)
+                                     , JSPChildSegment (..)
+                                     , JSPDescSegment (..)
+                                     , JSPSelector (..)
+                                     , JSPWildcardT (..)
+                                     )
 import Data.Functor                  (($>))
-import Data.Text                     (Text)
 import Data.Char                     (ord)
 import Text.ParserCombinators.Parsec ((<|>))
-import Language.Haskell.TH.Syntax    (Lift (..))
 
 import Prelude
-
-newtype JSPQuery
-  = JSPRoot [JSPSegment]
-  deriving (Eq, Show, Lift)
-
--- https://www.rfc-editor.org/rfc/rfc9535#name-segments-2
-data JSPSegment
-  = JSPChildSeg JSPChildSegment
-  | JSPDescSeg JSPDescSegment
-  deriving (Eq, Show, Lift)
-
--- https://www.rfc-editor.org/rfc/rfc9535#name-child-segment
-data JSPChildSegment
-  = JSPChildBracketed [JSPSelector]
-  | JSPChildMemberNameSH JSPNameSelector
-  | JSPChildWildSeg JSPWildcardT
-  deriving (Eq, Show, Lift)
-
-
--- https://www.rfc-editor.org/rfc/rfc9535#name-descendant-segment
-data JSPDescSegment
-  = JSPDescBracketed [JSPSelector]
-  | JSPDescMemberNameSH JSPNameSelector
-  | JSPDescWildSeg JSPWildcardT
-  deriving (Eq, Show, Lift)
-
--- https://www.rfc-editor.org/rfc/rfc9535#name-selectors-2
-data JSPSelector
-  = JSPNameSel JSPNameSelector
-  | JSPIndexSel JSPIndexSelector
-  | JSPSliceSel JSPSliceSelector
-  | JSPWildSel JSPWildcardT
-  deriving (Eq, Show, Lift)
-
-data JSPWildcardT = JSPWildcard
-  deriving (Eq, Show, Lift)
-
-type JSPNameSelector = Text
-
-type JSPIndexSelector = Int
-
-type JSPSliceSelector = (Maybe Int, Maybe Int, Int)
 
 pJSPQuery :: P.Parser JSPQuery
 pJSPQuery = do
